# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IIC Nepal API",
    "description": "REST API for Itahari International College — manages authentication, courses, and news/events.",
    "version": "1.0.0",
    "contact": {
      "name": "IIC Nepal",
      "url": "https://iic.edu.np"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current environment"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Admin login",
        "description": "Authenticate an admin user with email and password. Accounts with 2FA enabled receive a short-lived challenge and must call POST /auth/2fa/verify before a session is issued. Rate-limited to 5 attempts per 15 minutes per IP.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "example": {
                "email": "admin@iic.edu.np",
                "password": "securePassword123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, or a two-factor challenge is required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" },
                "example": { "requiresTwoFactor": true }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Invalid credentials" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Too many login attempts. Please try again after 15 minutes." }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/auth/2fa/verify": {
      "post": {
        "tags": ["Auth"],
        "summary": "Complete a two-factor login challenge",
        "description": "Verify a six-digit TOTP or one-time recovery code. Requires the challenge cookie created by POST /auth/login.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TwoFactorCodeRequest" },
              "example": { "code": "123456" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification successful and session cookie set",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "401": { "description": "Challenge expired or code invalid" },
          "429": { "description": "Rate limit exceeded" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/auth/2fa/setup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Start authenticator enrollment",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": {
                  "password": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Enrollment QR code and manual setup secret",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "qrCodeDataUrl": { "type": "string" },
                    "secret": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "409": { "description": "Two-factor authentication is already enabled" },
          "429": { "description": "Rate limit exceeded" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/auth/2fa/enable": {
      "post": {
        "tags": ["Auth"],
        "summary": "Confirm authenticator enrollment",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TwoFactorCodeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Two-factor authentication enabled and recovery codes issued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "recoveryCodes": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized or invalid code" },
          "409": { "description": "Setup has not started or 2FA is already enabled" },
          "429": { "description": "Rate limit exceeded" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/auth/2fa": {
      "delete": {
        "tags": ["Auth"],
        "summary": "Disable two-factor authentication",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DisableTwoFactorRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Two-factor authentication disabled",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized or invalid credentials" },
          "409": { "description": "Two-factor authentication is not enabled" },
          "429": { "description": "Rate limit exceeded" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie set by POST /auth/login. Use the 'Authorize' button after logging in; the browser will automatically send the cookie on subsequent requests."
      }
    },
...
    "responses": {
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Internal Server Error" }
          }
        }
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "requiresTwoFactor": { "type": "boolean" }
        },
        "required": ["requiresTwoFactor"]
      },
      "TwoFactorCodeRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Six-digit TOTP or a one-time recovery code"
          }
        }
      },
      "DisableTwoFactorRequest": {
        "type": "object",
        "required": ["password", "code"],
        "properties": {
          "password": { "type": "string", "format": "password" },
          "code": { "type": "string" }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
    }
  }
}